added chat schema to events

+++ server/src/controllers/event.js
@@ -4,7 +4,8 @@ const faker = require('faker')
 
 const Trail = require('../models/trail')
 const Event = require('../models/event')
-const User = require('../models/user.psql')
+const UserPsql = require('../models/user.psql')
+const UserMongo = require('../models/user.mongo')
 
 class EventController {
   async events(req, res) {
@@ -32,7 +33,9 @@ class EventController {
     const { eventId } = req.params
     const { fields } = req.query
 
-    const event = await Event.findById(eventId).lean()
+    const event = await Event.findById(eventId)
+      .select((fields && fields.replace(/,|;/g, ' ')) || '-chat')
+      .lean()
     res.json(event)
   }
 
@@ -73,7 +76,6 @@ class EventController {
       return res.status(400).json({ errors: errors.array() })
 
     const { id: userId } = req.context
-    // TODO: check if user owns the event
 
     const { eventId } = req.params
     const {
@@ -88,7 +90,7 @@ class EventController {
       return res.status(403).json(errorHandler('Nothing to update'))
 
     await Event.updateOne(
-      { _id: eventId },
+      { _id: eventId, userId },
       { title, description, date, duration_min, max_participants }
     )
 
@@ -98,9 +100,8 @@ class EventController {
 
   async deleteEvent(req, res) {
     const { id: userId } = req.context
-    // TODO: check if user owns the event
     const { eventId } = req.params
-    await Event.deleteOne({ _id: eventId })
+    await Event.deleteOne({ _id: eventId, userId })
     res.json({ success: true })
   }
 
@@ -114,7 +115,9 @@ class EventController {
     }).countDocuments()
 
     if (!joined) {
-      const imageUrl = faker.internet.avatar()
+      const user = await UserMongo.findOne({ _id: userId }).select(
+        'profileImage'
+      )
       await Event.updateOne(
         { _id: eventId },
         {
@@ -122,14 +125,14 @@ class EventController {
             participants: {
               userId,
               name,
-              imageUrl,
+              profileImage: user.profileImage,
             },
           },
         }
       )
     }
 
-    res.json({ success: true })
+    res.json({ success: !joined })
   }
 
   async leaveEvent(req, res) {
